+++ dev/interface/elevator_interface.h
@@ -50,7 +50,7 @@ public:
 
         time_msecs_t last_update_time;
 
-        int accmulate_angle;      // [degree]
+        float accmulate_angle; // [degree]
 
         float actual_velocity; // [degree/s]
